add site.miniGame functions to common/miniBoard
