+++ ui/simul/src/view/main.ts
@@ -5,6 +5,7 @@ import created from './created';
 import { richHTML } from 'common/richText';
 import results from './results';
 import pairings from './pairings';
+import { initMiniGames } from 'common/miniBoard';
 
 export default function(ctrl: SimulCtrl) {
   const handler = ctrl.data.isRunning ? started : ctrl.data.isFinished ? finished : created(showText);
@@ -19,7 +20,7 @@ export default function(ctrl: SimulCtrl) {
         }
       }),
     }),
-    h('div.simul__main.box', { hook: { postpatch: () => site.miniGame.initAll() } }, handler(ctrl)),
+    h('div.simul__main.box', { hook: { postpatch: () => initMiniGames() } }, handler(ctrl)),
     h('div.chat__members.none', { hook: onInsert(site.watchers) }),
   ]);
 }
