+++ ui/site/src/site.ts
@@ -10,7 +10,6 @@ import { unload, redirect, reload } from './reload';
 import announce from './announce';
 import { trans } from 'common/i18n';
 import sound from './sound';
-import * as miniGame from './miniGame';
 import { format as timeago, formatter as dateFormat, displayLocale } from './timeago';
 import watchers from './watchers';
 import { Chessground } from 'chessground';
@@ -33,7 +32,6 @@ s.watchers = watchers;
 s.announce = announce;
 s.trans = trans(s.siteI18n);
 s.sound = sound;
-s.miniGame = miniGame;
 s.timeago = timeago;
 s.dateFormat = dateFormat;
 s.displayLocale = displayLocale;
